pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes"
version = "0.1.0"
description = "Spacecraft trajectory calculator with gravity-assist optimization"
requires-python = ">=3.11"
dependencies = [
    "rebound>=4.0",
    "lamberthub>=0.1",
    "astropy>=6.0",
    "jplephem>=2.21",
    "scipy>=1.11",
    "numpy>=1.26",
    "plotly>=5.20",
    "matplotlib>=3.8",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "kaleido"]

[project.scripts]
hermes = "hermes.cli:main"

[tool.setuptools.packages.find]
include = ["hermes*"]
